DMRE-746

+++ common/Refresh_UAT.py
@@ -20,15 +20,15 @@ class SyncS3:
         if 'prod' in self._destination_path:
             print("error: please check the destination path. PROD keyword detected in destination path")
         else:
-            with open('logs/s3_sync_.txt', 'w+') as fout:
-                    print("Remove existing files in {0}".format(self._destination_path))
-                    command = "aws s3 rm {0} --recursive".format(self._destination_path)
-                    subprocess.run(command, shell=True, stdout=fout, stderr=fout)
+            if 'stage2' in self._source_path:
+                print("Remove existing files in {0}".format(self._destination_path))
+                command = "aws s3 rm {0} --recursive".format(self._destination_path)
+                subprocess.run(command, shell=True)
 
-                    print("Sync files in s3 from {0} to {1}".format(self._source_path, self._destination_path))
-                    command = "aws s3 sync {0} {1}".format(self._source_path, self._destination_path)
-                    subprocess.run(command, shell=True, stdout=fout, stderr=fout)
-                    print("completed")
+            print("Sync files in s3 from {0} to {1}".format(self._source_path, self._destination_path))
+            command = "aws s3 sync {0} {1}".format(self._source_path, self._destination_path)
+            subprocess.run(command, shell=True)
+            print("completed")
 
 
 if __name__ == "__main__":
